Feat - Thirty commit

+++ src/routes/memories.ts
@@ -1,14 +1,28 @@
 import { FastifyInstance } from "fastify"
 import { prisma } from "../lib/prisma"
 import { z } from "zod"
+import { request } from "http"
 
 export const memoriesRoutes = async (app: FastifyInstance) => {
-  app.get("/memories", async () => {
+
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get("/memories", async (request) => {
+    await request.jwtVerify()
+
+    const { sub: id } = request.user
+
     const memories = await prisma.memory.findMany({
       orderBy: {
         createdAt: "asc"
+      },
+      where: {
+        id
       }
     })
+
 
     return memories.map(memory => {
       return {
@@ -19,7 +33,7 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
     })
   })
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, replay) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -29,6 +43,10 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id }
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return replay.status(401).send()
+    }
 
     return memory
   })
@@ -47,7 +65,7 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
         content,
         coverUrl,
         isPublic,
-        userId: "56b9bee4-8094-4231-8ba9-65245eae9d84"
+        userId: request.user.sub
       }
     })
 
@@ -55,7 +73,7 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
   })
 
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, replay) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
@@ -70,6 +88,15 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
+    const foundedMemory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if (!foundedMemory.isPublic && foundedMemory.userId !== request.user.sub) {
+      return replay.status(401).send()
+    }
 
     const memory = await prisma.memory.update({
       where: { id },
@@ -85,12 +112,22 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
 
   })
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, replay) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return replay.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: { id }
